requirement.txt.txt
absl-py==0.12.0
asgiref==3.3.4
astor==0.8.1
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.6.20
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
contextlib2==0.6.0.post1
cycler==0.10.0
Cython==0.29.23
dataclasses==0.8
decorator==4.4.2
Django==3.2.4
enum34==1.1.10
ffmpeg==1.4
Flask==2.0.1
Flask-Cors==3.0.10
gast==0.2.2
gcloud==0.18.3
google-api-core==1.30.0
google-auth==1.30.2
google-cloud==0.34.0
google-cloud-core==1.6.0
google-cloud-speech==2.4.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
h5py==2.10.0
httplib2==0.19.1
idna==2.10
importlib-metadata==4.5.0
imutils==0.5.4
itsdangerous==2.0.1
jellyfish==0.8.2
Jinja2==3.0.1
joblib==1.0.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
libcst==0.3.19
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mock==4.0.3
more-itertools==8.8.0
mypy-extensions==0.4.3
mysql==0.0.3
mysql-connector-python==8.0.18
mysqlclient==2.0.3
networkx==2.5.1
nltk==3.6.2
nodejs==0.1.1
npm==0.1.1
numpy==1.16.1
oauth2client==4.1.3
opencv-python==4.5.2.54
opt-einsum==3.3.0
optional-django==0.1.0
packaging==20.9
pandas==1.1.5
Pillow==8.2.0
proto-plus==1.18.1
protobuf==3.17.3
protobuf-py3==2.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pycparser==2.20
pydub==0.25.1
pyparsing==2.4.7
python-dateutil==2.8.1
python3-protobuf==2.5.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.5.4
segtok==1.5.10
six==1.16.0
sqlparse==0.4.1
tabulate==0.8.9
tensorboard==1.15.0
tensorflow-cpu==1.15.0
tensorflow-cpu-estimator==1.15.1
tensorflow-estimator==1.15.1
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.8.1
torchvision==0.9.1
tqdm==4.61.0
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.5
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
yake==0.4.8
zipp==0.6.0
zope.interface==4.7.2